Fix compact icon missing right margin
In compact mode, the icon hits the right edge of the panel box (ugly!). Without compact mode, the label "NVPN" doesn't have a margin either, it just has a space appended. Unfortunately, it doesn't seem to be possible to fix this with CSS. So this fix just keeps the panel label always visible and toggles between the strings 'NVPN ' and ' ' (one blank space). Ugly, but works!