Minor UI label changes. (#35)
* Fixed location label comma placement

`_location_label.text` placed the comma next to the country, when it should have been next to the city.
How it was: `'* '+this.server_info.city+' ,'+this.server_info.country+' *'` -> `* City ,Country *`
How it is now: `'* '+this.server_info.city+', '+this.server_info.country+' *'` -> `* City, Country *`

* Minor UI label changes.

"Shown IP" isn't really indicative of what its representing. Changed to "External IP", which is more self-explanatory.
Capitalized "Uptime". The rest of the menu has capitalized descriptors, so it makes sense this should too.
Changed the semi-colan seperating the upload/download speed indicators to a pipe.
Upper-cased "tcp"/"udp".